refactor: type 별도분리
